Fix: Adapter Pattern 적용 #3
