+++ src/main/java/com/nhnacademy/marketgg/client/adapter/LabelAdapter.java
@@ -1,16 +1,15 @@
-package com.nhnacademy.marketgg.client.service;
+package com.nhnacademy.marketgg.client.adapter;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.nhnacademy.marketgg.client.dto.LabelRegisterRequest;
 import com.nhnacademy.marketgg.client.dto.LabelRetrieveResponse;
-
 import java.util.List;
 
-public interface LabelService {
+public interface LabelAdapter {
+
     void createLabel(final LabelRegisterRequest labelRequest) throws JsonProcessingException;
 
-    List<LabelRetrieveResponse> retrieveLabels();
+    List<LabelRetrieveResponse> retrieveResponse();
 
     void deleteLabel(final Long id);
-
 }
